fix: cleanup and renames
