--- backend_db/mock_ai_api.py
@@ -1,8 +1,0 @@
-# Temporary file
-
-import random as Random
-from typing import List
-
-
-def get_result(img_urls: List[str]) -> float:
-    return Random.uniform(0, 1)
