+++ backend_db/requirements.txt
@@ -2,7 +2,6 @@ flask
 flask-sqlalchemy
 flask-jwt-extended
 flask-bcrypt
-flask-cors
 psycopg2-binary
 python-dotenv
 pytest
