+++ backend_db/scrape_url_util.py
